refactor(confirmations): move useCustomAmountInfoStage to hooks/custom-amount
