+++ app/components/Views/confirmations/hooks/custom-amount/useCustomAmountInfoStage.test.ts
